requirements.txt
arrow==0.15.4
asgiref==3.2.3
binaryornot==0.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
coverage==5.0
Django==3.0
django-debug-toolbar==2.1
django-extensions==2.2.5
django-model-utils==4.0.0
future==0.18.2
idna==2.8
Jinja2==2.10.3
jinja2-time==0.2.0
MarkupSafe==1.1.1
pathlib==1.0.1
Pillow==6.2.1
poyo==0.5.0
psycopg2==2.8.4
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
style==1.1.6
update==0.0.1
urllib3==1.25.7
whichcraft==0.6.1
